Move HealthPack to mod.py
